requirements.txt
aider-chat==0.47.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
anthropic==0.28.1
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.2
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azure-storage-blob==12.20.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.34.153
botocore==1.34.153
Brotli==1.1.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chainlit==1.1.401
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
clarifai==10.5.4
clarifai-grpc==10.6.7
click==8.1.7
cohere==5.6.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
configparser==7.0.0
contextlib2==21.6.0
contourpy==1.2.1
crewai==0.35.8
crewai-tools==0.3.0
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.3
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dict-toolbox==5.0.0
diff-match-patch==20230430
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
duckduckgo_search==6.2.6
elevenlabs==1.6.1
embedchain==0.1.113
executing==2.0.1
fastapi==0.110.3
fastavro==1.9.5
filelock==3.15.4
filetype==1.2.0
flake8==7.1.0
FLAML==2.1.2
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-cloud-texttospeech==2.16.5
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gptcache==0.1.44
greenlet==3.0.3
grep-ast==0.3.2
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.1
importlib_resources==6.4.0
iniconfig==2.0.0
inquirerpy==0.3.4
instructor==1.3.3
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.1.20
langchain-anthropic==0.1.22
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.2.28
langchain-google-genai==1.0.8
langchain-groq==0.1.9
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.96
Lazify==0.4.0
lazy-object-proxy==1.10.0
litellm==1.41.27
literalai==0.0.607
llama-cloud==0.0.11
llama-index==0.10.59
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.59
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.11
llama-index-experimental==0.1.3
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.1.16
llama-index-llms-gemini==0.1.11
llama-index-llms-groq==0.1.4
llama-index-llms-openai==0.1.27
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.8
llama-index-postprocessor-cohere-rerank==0.1.7
llama-index-postprocessor-longllmlingua==0.1.2
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.32
llama-index-readers-llama-parse==0.1.6
llama-index-vector-stores-chroma==0.1.10
llama-parse==0.4.9
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.37.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
peewee==3.17.6
pexpect==4.9.0
pfzy==0.3.4
phidata==2.4.27
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
pop==27.1.0
pop-config==12.0.4
pop-loop==1.1.0
portalocker==2.10.1
posthog==3.5.0
PraisonAI==0.0.58
PraisonAI-Tools==0.0.8
primp==0.5.5
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.33
pycodestyle==2.12.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydeck==0.9.1
pyflakes==3.2.0
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.9.18
PyNaCl==1.5.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.374
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-rapidjson==1.19
python-socketio==5.11.3
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.5.15
rend==7.0.2
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
schema==0.7.5
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.23.1
semver==3.0.2
sentence-transformers==3.0.1
setuptools==72.1.0
shapely==2.0.5
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
stqdm==0.0.5
streamlit==1.36.0
striprtf==0.0.26
sympy==1.13.1
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.3
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trio==0.26.0
trio-websocket==0.11.1
tritonclient==2.48.0
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uptrace==1.26.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.25.0
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.20.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.41
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.19.2
